% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/fsrs.R
\name{fsrs_default_parameters}
\alias{fsrs_default_parameters}
\title{Get default FSRS parameters}
\usage{
fsrs_default_parameters()

fsrs_default_parameters()
}
\value{
Numeric vector of 19 parameters
}
\description{
Get default FSRS parameters

Get default FSRS parameters
}
\examples{
fsrs_default_parameters()
}
