% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/fsrs.R
\name{fsrs_retrievability}
\alias{fsrs_retrievability}
\title{Calculate retrievability (recall probability)}
\usage{
fsrs_retrievability(stability, elapsed_days)

fsrs_retrievability(stability, elapsed_days)
}
\arguments{
\item{stability}{Card stability}

\item{elapsed_days}{Days since last review}
}
\value{
Retrievability between 0 and 1
}
\description{
Calculate retrievability (recall probability)

Calculate probability of recall
}
\examples{
fsrs_retrievability(10.0, 5.0)
}
