% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fsrs_retrievability_vec}
\alias{fsrs_retrievability_vec}
\title{Vectorized retrievability calculation}
\usage{
fsrs_retrievability_vec(stability, elapsed_days)
}
\arguments{
\item{stability}{Vector of stability values}

\item{elapsed_days}{Vector of elapsed days (same length as stability)}
}
\description{
Vectorized retrievability calculation
}
