% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsrs_api.R
\name{fsrs_version}
\alias{fsrs_version}
\title{Get the FSRS algorithm version used by this package}
\usage{
fsrs_version()
}
\value{
Character string, e.g. "FSRS-5"
}
\description{
Get the FSRS algorithm version used by this package
}
