\name{fsrs_default_parameters}
\alias{fsrs_default_parameters}
\title{FSRS Default Parameters}
\description{
Returns the default 21 FSRS-5 parameters.
}
\usage{
fsrs_default_parameters()
}
\value{
Numeric vector of length 21 containing the default FSRS parameters.
}
\examples{
params <- fsrs_default_parameters()
length(params)
}
